edge cases lets go

+++ printf/ft_printf/ft_printf.h
@@ -7,8 +7,8 @@ typedef struct		s_node
 {
     char            *content;
 	size_t			i;
-	int				prec_len;
-	int				width_len;
+	long			prec_len;
+	long			width_len;
     bool            left_align;
     bool            has_prec;
     bool            has_width;
@@ -20,9 +20,9 @@ typedef struct		s_node
 }					t_node;
 
 char	*str_join_free(char **beg, char **end);
-void	handle_negative_wp(int *w_or_p_len, t_node *node);
+void	handle_negative_wp(long *w_or_p_len, t_node *node);
 char	*truncate_str(char *new_str, t_node *node);
-char	*get_filler(char *new_str, int len, bool pad_is_zero);
+char	*get_filler(char *new_str, long len, bool pad_is_zero);
 void	handle_zero(t_node *node);
 void	handle_asterisk(char *s, t_node *node, va_list ap);
 void	handle_dot(t_node *node);
@@ -34,7 +34,7 @@ void	add_letter(char s, char **to_print);
 int		manage_node( char *s, char **to_print, va_list ap, int i);
 int		iterate_string(char *s, char **to_print, va_list ap);
 int		ft_printf(const char *s, ...);
-void	update_padding(char *s, t_node *node, va_list ap, int *w_or_l, bool from_arg);
+void	update_padding(char *s, t_node *node, va_list ap, long *w_or_l, bool from_arg);
 void	handle_d(t_node *node, va_list ap);
